sys-power/powerstat: new ebuild
Powerstat measures the power consumption of a machine using the battery
stats or the Intel RAPL interface. The output is like vmstat but also
shows power consumption statistics. At the end of a run, powerstat will
calculate the average, standard deviation and min/max of the gathered data.

Package-Manager: portage-2.3.0