net-misc/curl: Backport fix muliple HTTPS conections
Gentoo-bug: 604104

net-misc/curl-7.52.1 is unable to sustain multiple
concurrent https connections with the openssl backend.
gnutls users are not affected.

This issue was introduced upstream in:
https://github.com/curl/curl/commit/cb4e2be7c6d42ca0780f8e0a747cecf9ba45f151

And resolved upstream in:
https://github.com/curl/curl/commit/a7b38c9dc98481e4a5fc37e51a8690337c674dfb

Because this version of curl is being stabilized as
part of  603574 / CVE-2016-9594 this shouldn't be left
until the next release.

Package-Manager: Portage-2.3.3, Repoman-2.3.1