dev-php/pecl-stomp: new revisions fixing the no-ssl build.
The main reason for these two new revisions are to fix what happens
when USE="-ssl" is set. The build system expects --with-openssl-dir=no
in that case, so the existing --without-openssl-dir was causing it to
crash. That should now be fixed; I've used "usex" instead of "use_with".

Some other minor fixes were made:

  * The LICENSE for this package is PHP-3.01, not PHP-3.
  * The "ssl" flag was also defaulted to being on, which -- unless
    critical -- kind of defeats the purpose of USE flags. It now
    defaults to being off.
  * Two eclass variables that were set to their default values have
    been removed.
  * The test suite is now enabled, but only for v2.0.0 where it actually
    passes (even though failure doesn't kill the build?).

Gentoo-Bug: 603560

Package-Manager: portage-2.3.0