dev-php/jpgraph: new version 4.0.2 fixing a few bugs.
This new version fixes a few outstanding issues with the old 3.x
series ebuilds. First, the config file is now installed to /etc, where
the user will be able to edit it, and then merely symlinked to
/usr/share/php/jpgraph where it used to live. This lets the user
configure caching at any time, which brings us to the second
improvement: caching is disabled by default, and we make no attempt
to configure the web server user or cache directory. This fixes an
open bug.

Finally, all of the documentation and examples are now installed in
the correct place, and no copies wind up in the PHP include directory.
Other minor improvements were made, like an update to EAPI=6, and the
use of a patch instead of a call to sed.

Gentoo-Bug: 520566

Package-Manager: portage-2.3.0