dev-php/phpdepend: new version 2.3.2 sans bundled dependencies.
Now that we have a framework for installing Composer packages, we
don't need to bundle the Symfony components that phpdepend uses. This
new version switches back to a source distribution and adds the
requisite RDEPENDs.

In addition, we now install phpdepend under /usr/share/phpdepend as
opposed to /usr/share/php/phpdepend. The latter is part of PHP's
"include" search path, and is only meant for library code.

Gentoo-Bug: 573340

Package-Manager: portage-2.3.0