dev-php/PEAR-Text_Highlighter: new revision sans the php-pear-r1 eclass.
This new revision drops php-pear-r1.eclass, and in the process fixes a
few bugs. We no longer install all of the development junk, which
eliminates the /usr/bin/generate executable (used to generate the PHP
classes), and avoids a file collision with app-accessibility/julius.
We also now store the documentation in the correct location.

One extraneous dependency was removed, and another undocumented one
added.

Gentoo-Bug: 544514

Package-Manager: portage-2.3.0