dev-db/firebird: don't use libintl.so SONAME for libfbintl.so
Firebird by default uses libintl.so.1 SONAME for libfbintl.so.1.
This means that ldconfig will make libintl.so.1 symlink to libfbintl.so.1.

However, actual libintl.so is a very different library, incompatible with
libfbintl.so.
This breaks packages which tries to link to it, notably
dev-perl/libintl-perl.

Fix it by making libfbintl.so SONAME match its actual name.

Package-Manager: Portage-2.3.3, Repoman-2.3.1