dev-php/twig: fix build with USE=extension unset.
In commit 14f2ab0, I accidentally broke building with
USE="-extension".  The eclass src_prepare phase was correctly skipped
(a new change), but then afterwards it became possible for the eclass
src_configure and src_compile to fail. Now those phases are skipped
when USE="-extension", too. For consistency and for a minor
performance improvement, the eclass src_unpack phase is skipped, too.

A new revision was made because I'm not 100% certain that it wasn't
possible to install the extension with USE="-extension" before. If
it was possible, any victims would want to re-emerge twig. I have
also added an unconditional RDEPEND on "dev-lang/php" that could
have been omitted before; that would itself justify a new revision.

Gentoo-Bug: 604874

Package-Manager: portage-2.3.0