dev-php/PEAR-Log: new version 1.13.1 replacing version 1.13.0.
This new version drops the php-pear-r1 eclass, and fixes a few
bugs. The documentation is now installed in the correct location, and
the examples are behind a USE flag. The "minimal" USE flag is gone,
because detection of features is automagic, and we don't want to
surpise users when they suddenly disappear. The test suite now also
works.

Package-Manager: portage-2.3.0