media-libs/harfbuzz: Raise glib minimum dependency for USE=glib
harfbuzz-1.3.3 raised this to 2.38 for G_TEST_DIST usage.
This was lowered again in 1.4.0 to support firefox and Windows use cases,
where glib 2.22 can still be used. But because G_TEST_DIST is still used
in the tests, keep this minimum version requirement in newer versions
as well. 1.4 now hava a fallback for G_TEST_DIST, but it is mentioned
that it then fails out of source tree tests. We have only newer than
2.38 glib versions anyways.

Package-Manager: Portage-2.3.3, Repoman-2.3.1