dev-php/agavi: new version 1.0.8.
This new version fixes a lot of problems caused by the old PEAR build
system. First, we don't need the PEAR dependencies, so they've been
dropped. The docs are installed in the correct place, and the build
system won't crash if you eselect the wrong version of PHP.

The ebuild has been updated to EAPI=6, and has a few more USE flags to
control its dependencies. Even the executable (which requires phing)
is now optional.

But beware: this package is mostly broken out-of-the-box. The "agavi"
executable requires some code from dev-php/phing, but the phing ebuild
was rewritten to use a PHAR, and no longer installs any source
code. Since the old version of phing (a) would not build for me and
(b) should have exactly the same problem, I don't consider that a
blocker for this new version.

Gentoo-Bug: 604954

Package-Manager: portage-2.3.0