dev-php/phing: new unbundled version 2.16.0.
This new version is a major overhaul of the ebuild. We start with an
update to EAPI=6, and then everything else is basically new. The
previous version installed a single compressed PHAR archive, while now
we actually install the source, documentation, examples, and
executables -- all in the correct places.

Some dependencies were fixed, and there are new USE flags for a couple
of optional dependencies (eventually there should be more flags,
rather than rely on USE=minimal). Upstream's metadata was updated,
too.

Unfortunately, many features of this package will not work. Since
phpDocumentor and phpunit are still packaged as PHAR archives, phing
cannot make use of their libraries. This is not a new problem,
however, so it is acceptable to allow in in the new version.

Gentoo-Bug: 604954

Package-Manager: portage-2.3.0