net-dns/rbldnsd: new revision to fix an out-of-memory error.
This new revision adds a patch that has already been merged (but not
released) by upstream. It should fix an out-of-memory error that
occurs when trying to load a huge zone.

Additionally, upstream had undergone a non-hostile takeover by the
Spamhaus project, so the HOMEPAGE and SRC_URI have been updated.

Gentoo-Bug: 603330

Package-Manager: portage-2.3.0