x11-terms/xterm: needs x11-apps/rgb for colors
This package needs to runtime depend on x11-apps/rgb, because without it
installed there are no colors displayed in an xterm window - besides black.
Text in any other color is not shown.

x11-apps/rgb is currently pulled in by x11-base/xorg-server, but there are
also ways to run xterm on a system which do not use the Xorg server: for
example NX.

Gentoo-Bug: https://bugs.gentoo.org/545496
Package-Manager: Portage-2.3.3, Repoman-2.3.1