dev-php/PEAR-PHP_Beautifier: new revision sans the php-pear-r1 eclass.
This new revision drops the php-pear-r1 eclass in order to fix some
otherwise-unfixable issues. First, the ebuild has been updated to
EAPI=6; that one was easy. Two new USE flags to enable the CLI and
installation of the examples were added -- that and the removal
of the eclass caused some RDEPEND changes.

We now have custom src_prepare() and src_install() phases:

  * In src_prepare, we replace the strings @package_version@ and
    @php_bin@ everywhere in the source code using sed. We also
    apply a new patch to use the default error_reporting, which
    hides a bunch of ugly errors.

  * The src_install phase now installs the examples to the proper
    location, and only installs the CLI script with USE=cli.

Package-Manager: portage-2.3.0