dev-php/PEAR-Net_DNSBL: new EAPI=6 revision.
This new revision drops the php-pear-r1 eclass, which means that the
examples are now installed in the correct place. They're also hidden
behind USE=examples, as are the associated dependencies.

The other major reason for the new revision is to drop the erroneous
RDEPEND on dev-php/PEAR-Net_CheckIP. That package is not needed, and
should be last-rited once this new revision has stabilized.

Package-Manager: portage-2.3.0