net-dns/updatedd: new revision to set timeouts for ipserv.pl.
This new revision adds a patch by Maciej S. Szmigiero that sets
timeouts on the socket used by ipserv.pl. Doing so prevents updatedd
from hanging when the other end of the socket goes AWOL.

As a consequence, dev-perl/IO-Socket-Timeout was added to RDEPEND. The
"~arm" and "~ppc" keywords had to be dropped to satisfy that new
dependency.

Gentoo-Bug: 605626
Package-Manager: portage-2.3.0