x11-libs/pango: Clarify what USE=X does in local description
pangoxft is more or less a legacy library, useful for libXaw/libXft
or other pure X11 using applications that want to integrate pango
with the existing libXft stuff. Most modern things do not use this,
so clarify this in a local USE flag description in metadata.xml.
Many consumers mistakenly USE depend on this without actually looking
for or using pangoxft, so this also hopefully avoids further usage
of this.
The goal is to get rid of the unnecessary USE deps on consumers, where
it is indeed unnecessary, and then perhaps rename the USE flag from
IUSE=X to IUSE=xft or some such, so that it's not built by default
on desktop profiles.

Package-Manager: Portage-2.3.3, Repoman-2.3.1