dev-perl/mime-construct: add "Date" field to generated messages
Currently, mime-construct does not include the "Date" field in generated
messages.

The "Date" field is required according to the RFC 5322 and not every MTA or
SMTP client adds it when it is missing (for example mail-mta/msmtp does not
do it).

Mail clients often timestamp messages lacking this field as either
01.01.1970 or a date the message was first seen by a particular client
rather than as an actual posting date.

This issue was reported upstream a year ago but there was no response - the
software had last release in 2010 so it is probably no longer maintained.

That's why we have to add the missing field ourselves.

Gentoo-Bug: https://bugs.gentoo.org/583882
Package-Manager: Portage-2.3.3, Repoman-2.3.1
Closes: https://github.com/gentoo/gentoo/pull/3408