x11-libs/gdk-pixbuf: Clarify what USE=X does in local description
gdk_pixbuf_xlib is only useful for pure X11 applications, GTK+ does
not use this.
The goal is to get rid of the unnecessary USE deps on consumers, where
it is indeed unnecessary, and then perhaps rename the USE flag from
IUSE=X to IUSE=xlib or some such, so that it's not built by default
on desktop profiles. Though w3m is a consumer and might be in default
deptrees via xdg-utils

Package-Manager: Portage-2.3.3, Repoman-2.3.1