dev-php/PEAR-Console_CommandLine: new version 1.2.2.
This new version drops the php-pear-r1 eclass in exchange for a few
fixes. The documentation (examples) are now installed in the correct
location, behind USE=examples. The xmlschema.rng data file is
installed outside of the PHP include directory, since it is not
included by PHP code. And finally, the (working) test suite has been
enabled.

Package-Manager: portage-2.3.0