dev-lang/php: new security releases for all three series.
In addition to the version bumps, the (unused, hopefully) libtool
eclass was dropped from the php-5.6.30 and php-7.0.15 ebuilds
Upstream merged a fix for heimdal detection, allowing us to drop that
workaround from src_prepare. After the heimdal workaround was removed,
it then made sense to drop the autotools eclass entirely. That means
no more custom branding in the 7.0 series, but we are essentially a
vanilla upstream PHP at this point.

Gentoo-Bug: 597448
Gentoo-Bug: 597450

Package-Manager: portage-2.3.0