dev-lang/perl,virtual/perl-*: Cleanup all non-latest perl-5.22* and perl-5.24* re bug #589680
Keeping 5.20.0 for now while backwards deps are still unsatisfied,
simply cleaning out all but the latest in each ABI subslot to
minimise the number of people who are going to be forced to do an
ABI rebuild prematurely.

If you still need perl-5.22.0, 5.22.1, 5.22.0, 5.24.0 , please
locally revert this commit.

Note: You may also need a copy of files/ at the time this commit
was performed, as those packages use horrible old eblits, which
we're trying to kill off

Bug: https://bugs.gentoo.org/589680
Bug: https://bugs.gentoo.org/586418

Package-Manager: Portage-2.3.3, Repoman-2.3.1