sys-libs/compiler-rt: Add slotted version
Introduce a slotted variant of compiler-rt libraries with the slot
matching the clang version. This will make it possible to install a new
compiler-rt version before upgrading clang to the corresponding version,
therefore preserving a working compiler through (even minor) clang
upgrades.

The alternative was to replace the complete version number with the
major version in the runtime directory path. However, software (e.g.
Mesa) hardcodes the default path and therefore breaks when it is
changed.