sys-kernel/ck-sources: version bump (4.9.6-r1 = genpatches-4.9-8)
The issue described by kernel.org commit history as follows:

[...] Optimus/nouveau will fail in a mysterious way if the delayed event is handled as soon
as possible like it is done in drm_helper_probe_single_connector_modes() in case the poll was
enabled before. [...] Adding 1sec delay to the poll_work is enough to work around the issue
in Optimus setups and gives shorter response on handling the initial delayed event.

This is fixed by genpatches-4.9-8

Package-Manager: portage-2.3.3