sys-apps/policycoreutils: Properly depend on audit[python] and fix selocal attribute resolving
The first bug this release fixes is bug 597978. Previous policycoreutils versions
provided a selocal application which can't deal with the new setools. The previous
setools had seinfo return a non-zero return code if an attribute (or role, or ...
depending on what was asked) was not found. The newer setools gives this in the
output, but keeps the zero return code. selocal depended on the return code
previously.

The second bug this release fixes is bug 605692. Well, partially, because the bug
has two issues mentioned. The first one is the (wrong) dependency on audit. If the
policycoreutils package is build with USE="audit" then it should depend on audit[python]
rather than just audit. This is the issue that is fixed in this release.

The second issue (not being able to use a number of semanage commands when
USE="-audit") is not resolved yet.

Package-Manager: portage-2.3.3