mail-filter/spamassassin: disable interactive prompts properly.
The PERL_MM_USE_DEFAULT variable is used to disable the interactive
prompts in SpamAssassin's build system. However, when I dropped the
perl-module.eclass a while back, I only set the variable where it
seemed to be necessary. Where it is necessary turns out to depend on
the package manager, because different package managers may treat the
controlling terminal and file descriptors differently, and that's how
MakeMaker decides whether or not to prompt you.

That problem manifests in Paludis, and is hopefully fixed by exporting
PERL_MM_USE_DEFAULT=1 sooner, in src_configure(). This change should
only affect the build, so it is made without a new revision.

Gentoo-Bug: 604518

Package-Manager: portage-2.3.0