sys-libs/efivar: add workaround for GCC_SPECS
Lars Wendler reported a link failure when using a forced as-needed
gcc profile.

This failure was caused by gcc ignoring the specs file passed on the command
line, in favor of the GCC_SPECS environment variable.

To work around this, we append the necessary specs file to the envvar.

Package-Manager: Portage-2.3.3_p56, Repoman-2.3.1_p49