virtual/imap-c-client: new version with a "kerberos" USE flag.
At least one of our consumers (dev-lang/php) needs to have its
kerberos support match the c-client implementation. In order to
propagate the USE=kerberos dependency into the implementation,
the virtual needs to have it too. So this commit adds it.

Thanks are due to Dennis Schridde who both reported the problem and
supplied a fix.

Gentoo-Bug: 541046

Package-Manager: portage-2.3.3