dev-lang/php: two new 7.x versions and a bugfix.
Two new versions 7.0.16 and 7.1.2 were released recently and this
commit adds them to the tree (nothing out of the ordinary there).

In addition, a bugfix was made to both new 7.x versions as well as a
new 5.6.30-r1 revision. The kerberos support in dev-lang/php must
match that of its c-client implementation when USE=imap is set. In the
past, that was impossible to enforce, because virtual/imap-c-client
was missing the "kerberos" flag. Now that it has been added, the
additional dependency has been added in PHP. Thanks to Dennis Schridde
for noticing and fixing that bug.

Gentoo-Bug: 541046

Package-Manager: portage-2.3.3