python-single-r1.eclass: Always output use-cond in PYTHON_DEPS
Output a USE conditional in PYTHON_DEPS even when using only a single
target. While this does not make a strict difference since the target
flag is in REQUIRED_USE anyway, it results in more unified output
and makes it easier to catch mistakes if someone plays with the eclass
in the future.