Add xorg-server-1.19 support to tigervnc-1.7.1
The latest release of tigervnc (1.7.1) doesn't
support xorg-server-1.19. Backporting a patch
and commit from the main trunk allows one to
build tigervnc-1.7.1 with the server use flag
for xorg-server-1.19.

The ebuild is tested to be building and performing
correctly with LXDE and GNOME on a localhost.
However, I recommend backporting other changes as
well to fix possible issues with xorg-server-1.19.

Since TigerVNC releases aren't as frequent,
backporting might be worth the effort.

Commited by: Bruno Henc <hencb @ protonmail . ch>