net-libs/c-client: use ":" as pattern separator in a sed call.
Most of the "sed" calls in the c-client ebuilds use ":" as the pattern
separator, because we expect to be replacing things with front-slashes
in them. One exception was the call that replaced CC, AR, and RANLIB;
and that finally bit us. A user with absolute paths in those variables
reported the defect.

This commit fixes the call to use ":" as the pattern separator, like
everywhere else in the ebuild. Since it is a build-only fix, it was
made without a new revision.

Gentoo-Bug: 609744

Package-Manager: portage-2.3.3