app-emulation/runc: apply fix for CVE-2016-8867
Resolves bug #598364, but for real this time.

Docker uses a commit of runc (50a19c6) which is not tagged in
opencontainers/runc, and so pulling the official rc2 tarball from
opencontainers/runc did not pick up this fix.

When docker-1.12.3 was released, the reason it fixed the above CVE upstream was
that it built runc from such a commit (specifically from this branch
https://github.com/docker/runc/commits/1.12.x).

Because gentoo's runc ebuild pulls from the proper upstream, it did not pickup
that fix when updating to 1.12.3.

This takes the patch applied in docker's 1.12.x branch of runc and applies it.

Package-Manager: portage-2.3.3