ruby-ng.eclass: Replace unnecessary 'eval ls' with array fnexp
Replace the unnecessary use of 'eval ls -d ...' with much simpler
and safer filename expansion via bash array. The 'eval' was completely
unnecessary in the original code; however, the late addition of quoting
would have broken it if eval did not implicitly discard the quotes.
The 'ls -d' was unnecessary as well since bash performs filename
expansion before passing the parameter to 'ls'.

Furthermore, a check for accidental multiple expansion has been added.
A complementary check for failed expansion can not be added since
the function is called in src_unpack() as well and the expansion fails
then.