dev-libs/msgpack: version bump to 1.4.2 and 2.2.1
An update to the latest version of 1.X and 2.X of the package. This was
already being worked on in PR #2974, I processed the comments and added
some extra things:

* Updated copyright
* EAPI version bump
* Updated license to Boost-1.0
* Added examples USE-flag
* Added doc USE-flag
* Updated patches for cflags and static-libs
* Masked for testing

The new versions are masked for testing as the status of depending
packages is currently unknown. Bugs for the package maintainers will be
created.

Gentoo-Bug: https://bugs.gentoo.org/show_bug.cgi?id=601370

Package-Manager: Portage-2.3.3, Repoman-2.3.1