dev-lang/spidermonkey:{38,45} ebuild improvements and bugfixes
Upstream stripped out most slotting and also was installing headers
as symlinks.  This has been addressed in :38 through patches from GNOME.

Redundant and unused eclasses were removed, which should help address
bug 611232.

spidermonkey-45 still installs headers as symlinks, and I have ben unable
to figure out where in the python buildsystem code this is being triggered
to reverse it back to a copy; the ebuild is now p.masked because of that.

Package-Manager: portage-2.3.3