net-dialup/minicom: Force lockdir to /var/lock
Upstream uses automagic detection of the first existing directory of:
/etc/locks /var/lock /usr/spool/locks /var/spool/locks /var/spool/lock
/usr/spool/uucp /var/spool/uucp /var/run

While many of these will exist in a live system, that was not a safe
assumption during a chrooted package build, such as stage building.

The init system is responsible for creating /var/lock, so force the
lockdir to that.

Package-Manager: portage-2.3.3
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>