ltprune.eclass: Split prune_libtool_files out of eutils
Move the prune_libtool_files into a dedicated ltprune.eclass.
The function is quite complex and depends on toolchain-funcs. It has
a separate maintainer, is not useful to non-autotools ebuilds, and even
there it is frequently replaced by the simpler 'find ... -delete' call.

Kill the unnecessary Prefix logic. There is no functional difference
between starting the find in ${D} and ${ED}, and the latter implies
unnecessary hackery for old EAPIs.

The new eclass is implicitly inherited by eutils in EAPI 6 and older
in order to preserve compatibility with existing ebuilds. However, all
ebuilds should switch to inheriting it directly.

The split has been suggested by Ulrich Müller.