mail-client/thunderbird: fix enigmail symlink installation
Limitations in package managers prevent a symbolic link from replacing a directory
at merge time.  This patch addresses that issue by removing the directory belonging
to the currently installed thunderbird in pkg_preinst().  It's been tested against
a direct upgrade form 45.7.0 as well as a reinstall, and seems to do the 'right thing'
in all cases.

Package-Manager: portage-2.3.3