app-portage/eclass-manpages: Clearly indicate the live ebuild, #612422
The current eclass-manpages ebuild uses the live tree to generate
the manpages. The resulting manpages can therefore differ on each
rebuild, and the ebuild can trigger broken output at random points
in time depending on the current state of eclasses. Furthermore,
the package contents randomly become stale and the user needs to rebuild
it manually in order to obtain new contents.

All this considered, the current ebuild is in essence a live ebuild.
Following the policies for live ebuilds, rename it to version 9999*
and remove all the keywords. Users wishing to use the live ebuild can
unmask it per the policy.