net-libs/libgfbgraph is missing use flag on deps
Introspection use flag is required to compile
libgfbgraph with glib-json and rest even if
the introspection use flag is disable on
libgfbgraph.

This issue only appears if introspection isn't
set globally to enabled. If GNOME packages don't
follow the convention of listing all use flags
in their dependencies and rely on a global use
flag you can simply ignore this change.

Issue tracked here:
https://bugs.funtoo.org/browse/FL-3647
No equivalent issue was found on the Gentoo
bugtracker.