gnome-extra/evolution-data-server: fix Google Contacts/Tasks support with USE="gnome-online-accounts -google"
--enable-google controls whether evolution knows how to talk to Google outside IMAP/SMTP/CalDAV; this should
have been named --enable-gdata by upstream (and might get renamed to that soon, actually).
--enable-goa controls whether gnome-online-accounts is supported as a credentials backend for Google OAuth2.
--enable-uoa controls whether ubuntu-online-accounts is supported as a credentials backend for Google OAuth2;
we do not support this (at least currently).
--enable-google-auth controls whether an internal Google support for OAuth2 credentials is built and supported.

Rework the logic based on this to just always enable Contacts/Tasks support when either way to
support Google OAuth2 is enabled and thus available. Both can be enabled as well and is suggested
by upstream to do so. Also clarify metadata.xml with this, and remove the security warnings as
3.22 is not affected by this (having ported to SLOT=4) and the impact to 3.20 is minimal as only
Google OAuth2 login/auth view is shown by it.

Package-Manager: Portage-2.3.4, Repoman-2.3.1