net-libs/rb_libtorrent: Make sure 1.1.2 links against the right version of the library
Also makes sure that we are using the right header files. When building the
python bindings, since we are working in a copy of the source tree, rather
than the one the lib was built in, it tends to use the installed versions
rather the new ones.

Package-Manager: Portage-2.3.4, Repoman-2.3.2