epunt-cxx.eclass: Split C++ check punting code out of eutils
Split the epunt_cxx (plus internal code) to a dedicated eclass. This is
rarely needed, usually indicates a dead upstream and requires
the ELT-patches framework. The patches are going to be split to
a separate package, and the new eclass will therefore need to DEPEND
on it. We do not want the dependency to apply to all eutils users
though.