dev-php/pecl-http: new revisions to fix extension load order.
The http extension has dependencies on two other extensions, and needs
to be loaded after them. Our eclass currently can't handle that -- we
throw files into /etc/php/.../ext-active, and they get loaded in glob
order. For an ugly workaround, I've renamed the symlinks that get
installed in ext-active to "zzhttp.ini", but have left them pointing
to a file named "http.ini" in the "ext" directory.

That fixes the problem for pecl-http, but it would be nice to have a
long-term solution for the problem. I've made a note to that effect on
bug 586446, where we are tracking future improvements to the eclass.

Gentoo-Bug: 586446
Gentoo-Bug: 612054

Package-Manager: Portage-2.3.3, Repoman-2.3.1