media-gfx/rawtherapee: version bump
New version, with a lot of new features and a switch from gtk2 to gtk3.
Thanks Franz Fellner for converting the ebuild fromn git to tarballs
and Marek Szuba (marecki) for suggestion on how to improve it.

See bug #606896 for details, #c13 for Franz ebuild and #c17 for Marek
suggestions.

Signed-off-by: Luca Barbato <lu_zero@gentoo.org>