gnome-extra/evolution-data-server: bump to 3.22.7; re-enables POP extensions for new POP accounts
This was done to 3.22.6 upstream to avoid issues with Yahoo mail (by my reading slowness
due to PIPELINING extension), but many basic POP3 functionality is via extensions
(UIDL for duplicate mail detection, extended error codes, etc), so this bump just
re-enables them again in the form of new account default setting.
This is the only change, so without POP3 usage there's no hurry to upgrade, however this
is also a stable candidate, and we don't want to stabilize the previous version that has
this issue. Welcome to ~arch.