www-client/firefox: add nsplugin use flag to override the NPAPI-plugin-disable default
As of 52.0, mozilla has dropped support of all NPAPI plugins except for flash.  52.0ESR
(which is what gentoo stabilized) will retain this support, but for this particular
package we can also hotwire the switch used in 52.0ESR to allow plugin support to continue
on the release version.  Note, this may not work after 52 and if so the flag will be
removed.

Package-Manager: portage-2.3.3