dev-lang/php: new revisions to phase out PHP_INI_VERSION.
These three new revisions install the "production" version of php.ini
unconditionally. The changes between the two versions are minimal
anway, and the production version is safer for people who don't care.

A warning will be emitted for users who still have the variable set.

In addition, the 5.6 and 7.0 series will now cache their ./configure
test results. This is based on a similar change made in 7.1, and
should speed up the build a little.

Gentoo-Bug: 530002
Gentoo-Bug: 611214

Package-Manager: Portage-2.3.3, Repoman-2.3.1