media-libs/grilo: bump to 0.3.3, fix up test dependencies
Upstream commented out the python tests a while ago in 2015 via an
exit(0) call in testrunner.py, but we kept depending on pygobject both
slots (for gi.repository AND "import glib" usage in the python tests).
Remove the unnecessary deps and replace it with only any python provider,
as tests still want to launch testrunner.py via $PYTHON, for which we
want to set up the variable still, but all the pygobject stuff can go away
for now.
The python code before exit(0) should work with any python interpreter
version, so this extends it to accept all cpython implementations too.