media-gfx/gnome-photos: bump to 3.22.5; rework grilo-plugins handling, add missing online miners dep
gnome-photos optionally depends on upnp-av and flickr out of grilo-plugins only, and we were
missing flickr. Instead of hard depending on flickr, in addition to upnp-av, move them both
under a USE flag, so those not using flickr or DLNA don't get forced to pull in the whole
dleyna stack or the flickr stuff (though the latter is just oauth).

Similarly for pulling photos, gnome-online-miners are used, but this dependency wasn't specified.
So add it as a RDEPEND and move grilo-plugins as a pure RDEPEND as well. gnome-online-miners
always has facebook/google/media-server support, which gnome-photos uses, but it also optionally
uses flickr online miner, so echo the USE dependency to that as well.