dev-lang/perl: Add 5.26.9999 periodic-snapshot version
This is a pre-release of 5.26 series and aims to be a
work-in-progress for the life of Perl 5.26.

Changes are intended to happen here first, and then get folded
out into actual releases.

This version requires grown-up mode and likely judicious use
of package.provided ( because at least at this time, requisite
virtuals are missing )

Also, this package is very ambitious and employs INC modifications
that incorporate a broader selection of previous perls' versions,
by re-including the architecture-independent paths from older
perls.

Its not known at this time if this is a viable strategy for a
final release, but it intends to make testing this easier.

( Because you might need lots of --nodeps to use this version
  to begin with )

Also, you may want to borrow from
  https://github.com/gentoo-perl/perl-testing-profiles

Which has a few things optimised for doing this, and for testing
in general, including some workarounds for known defects / spurious
failures.

Package-Manager: Portage-2.3.4, Repoman-2.3.2