net-mail/courier-imap: new EAPI=6 version to fix a few bugs.
This new version contains upstream fixes for the ongoing removal of
the SSL3 protocol. That should fix bug 593064, which otherwise would
require SSL3 support in OpenSSL. The new version also renames
"maildirmake" (and its man page) to "courier-maildirmake" to avoid a
collision with mail-mta/netqmail (bug 482098).

During the upgrade, the ebuild was updated to use EAPI=6. The patches
it uses were bumped to be -p1 compatible, and now the PATCHES array
and eapply are used in src_prepare. The eutils and multilib eclasses
were dropped, since they were only used for epatch and get_libdir.

I also tried to clean up a few bits in src_install. I changed one use
of $WORKDIR/tmp to $T, and tried to avoid changing directories so much
(by spelling out the target paths). I removed a recursive dohtml, and
replaced it with a few calls to dodoc. I also fixed a bare "chmod" to
use fperms.

Gentoo-Bug: 482098
Gentoo-Bug: 593064

Package-Manager: Portage-2.3.3, Repoman-2.3.1