media-sound/picard: make chromaprint and python-discid runtime optional
Optional runtime dependencies are:
* chromaprint[tools]
* python-discid

Refer to https://devmanual.gentoo.org/general-concepts/use-flags/ for
information on why these dependencies should not be listed in DEPEND
and/or RDEPEND. Picard is capable of executing correctly with and
without these dependencies installed on the system. A user can install
these dependencies at a later stage and Picard will recognise them
without a need to rebuild the media-sound/picard package.

Package-Manager: Portage-2.3.5, Repoman-2.3.2