dev-libs/glib: fix GIO module cache file ownership and cross-compilation
Bug 518422 tried to fix module cache file ownership to glib, but the ebuild
changes missed /modules path and used $libdir/gio/ instead of $libdir/gio/modules

While cross-compiling, we can't generate the cache file, so we shouldn't
die out either. Instead of removing the die, rework it all to not do any
of it while cross-compiling, because we'd end up with an empty glib owned
cache file, which could result in no modules working at all as a cache
file is present from the preinst touch.

Gentoo-bug: 518422
Gentoo-bug: 613306
Thanks-to: Alexis Ballier <aballier@gentoo.org>