dev-libs/glib: Export PCRE_LIBS in the no-pkgconfig case
We have by now started to use the system libpcre unconditionally,
instead of the bundled version, but this needs pkgconfig to be
found. So we need to handle this in the no-pkgconfig case as well,
alongside FFI. However this suggests there is no point in this
case anymore and we should just build depend on pkgconfig, as
we can't get a working pkgconfig anymore without bootstrapping it
via USE=internal-glib as libpcre currently needs pkgconfig as
a build dependency already. Until this is still in the pondering
phase, this PCRE_LIBS exporting should at least fix the cases
where pkgconfig got depcleaned inbetween but libpcre exists on
system.

Gentoo-bug: 615092