glsa.dtd: Allow slot="" attribute for vulnerable&unaffected
Officially list the slot="" attribute that is used in GLSAs for quite
some time in the DTD. It is supported by Portage and gentoolkit for
a long time, and was used in GLSAs interchangeably with implicit
appended ':slot' to the version. However, the latter was ugly and worked
only by accident, so we are moving towards the former.