app-eselect/eselect-postgresql: Bug Fixes and Enhancements
Version 2.0 is capable of handling the upcoming change in versioning
for PostgreSQL.

Links to pkg-config files of the selected slot are now created and
managed by this module.

Links to the man pages for the selected slot are now created and
managed by this module. Slot-specific man pages are handled by the
ebuilds.

This version is stateless. It does not need to store any files to
determine which slots and links are in use.

Bugs: 597564, 512236, 564512

Package-Manager: portage-2.3.0