www-client/firefox-bin: add apulse compatibility to wrapper and some use-flags
Added ffmpeg and pulseaudio to IUSE, default-enabled, to be (A) compatible with
www-client/firefox and (B) ensure dependencies for this functionality is brought
in and retained without @world entries required.  It's a binary package so the
hit is minimal to be dealing with optional runtime-only deps with use flags.

Also added apulse to the LD_LIBRARY_PATH so that it will be used to provide
audio support when it is installed.

Bug: http://bugs.gentoo.org/615000

Package-Manager: Portage-2.3.3, Repoman-2.3.1