www-client/firefox: bump to 52.0.2, fix compilation on musl, allow apulse
This bump also acts to trigger rebuilds for the *DEPEND changes that
allow apulse to be used at runtime instead of just pulseaudio. Note that
apulse support is not a drop-in replacement and end-users need to do something
to ensure the apulse version of libpulse*.so are in the LD_LIBRARY_PATH when
firefox is run.  The mozilla-overlay contains an attempt at ensuring these libs
load automatically but this will not be added to the gentoo repo until it is
ensured that there are no negative results or conflicts as a result.  End-users
should continue to use alsa if the do not wish to use the full version of
pulseaudio.

Bug: http://bugs.gentoo.org/613980
Bug: http://bugs.gentoo.org/615372

Package-Manager: Portage-2.3.3, Repoman-2.3.1